Fix storage dev_file mode to be in public/ dir
So it can be served as a static file. This is _really_ not suitable for production -- zero access controls to access -- but it's good for testing, for now.

Also keep dev_file mode files in directories named by Rails.env, so we can clear out test without clearing out dev files