import updated_at
In order to do this, we had to figure out how to tell ActiveRecord to turn off auto-setting of updated_at, so our relationship and other cleanup code wouldn't re-set it to something other than what we wanted from the update.

This means our importers are now wholly responsbible for setting updated_at and created_at manually, and if they don't, the db constraint will complain.

So the importers now set to DateTime.now if they didn't find one in the import json. AND they complain about not finding one in the import.json

Right now our fileset and collection json is missing the dates, so it will result in a lot of complaints. See #92. Maybe fix that before merging this? Not sure.