Blacklight routes
As generated by BL generator, with editing to fit it into what we've got.

We are trying to ONLY add Catalog routes, since that's the only part of BL we meant to use. So far mostly working.

At the moment what's NOT working is BL default layout assumes it can have a link to devise registerable routes, which we aren't using. If there is a logged in user. We'll have to clean that up momentarily.