Fix chromedriver 74.0's failure to scroll to reveal something covered by a position:fixed element
As of Chrome/chromedriver 74.0, chromedriver will refuse to click on something
if it's covered up by something with "position: sticky", even if scrolling
could uncover it. Chromedriver will normally scroll to reveal something to click
on it, and chromedriver 73.0 handled it fine.

I think this is probably a chromedriver bug, but don't know how/where
to report it. For now, this is a workaround, we occasionally hackily scroll
to top so chromedriver can find whatever we wanted to click on, at random places
where it was failing otherwise.