Give ViewModel a valid_model_type_names macro
to specify what classes are allowed for the first arg to the ViewModel initializer. Fail early to avoid confusing bugs later, since view models are generally intended to handle a particular kind of 'model'.

Abstract out the code that was previously custom in the ResultThumbDisplay, and use the new valid_model_type_names macro in existing view models.