move member image presentation css to work_show.scss
Not sure if these two different scss files, to separate out 'layout' CSS are justified. They are kind of interrelated. But we'll leave them separate for now. The idea was that something other than a work show page might want to use the layout CSS for our kind of crazy two-columns with sophisticated repsonsive behavior.