update some indirect yarn dependencies
To avoid vulnerable versions as flagged by github

WS-2019-0032
CVE-2019-10744
CVE-2019-10744

It seems like there ought to be a way to do this without editing the package.json, but oddly it appears there is not.
https://github.com/yarnpkg/yarn/issues/4986

So one way to do this is to use 'resolutions' like this. We add a resolution that (unlike the 'resolution' doc examples at https://yarnpkg.com/lang/en/docs/selective-version-resolutions/) is global (NOT scoped to a particular dependency), and that simply says we insist on versions being recent enough to avoid reported vulnerable versions.

This ought to avoid locking the resolution any MORE than we want, potentially conflicting with future indirect dependency requirements, doing no more than what we want, insisting it's at least the patched version(s).

Then we just run "yarn install" again to update the yarn.lock.

Another workaround way to do this mentioned in the yarn GH issue would be manually removing the relevant section of the yarn.lock, then running "yarn install" again to force yarn to find latest resolvable version of the indirect dependencies we wanted to update.