don't use webpacker externals for Uppy
Just count on it being there, without an import statement. Webpacker seems to let us get away with it.

The explicit externals/import statement was a problem because we only load uppy from CDN on admin layout. To do that for real in webpacker (conditionally load it only on admin), we'd need separate packs, which we don't have yet.