Only create DZI for image files
We still run DZI deletion routines for non-images on asset destroy/update. If they don't
find any existing DZI files to delete, they ought not to complain. This is simpler than
trying to figure out exactly if DZI deletion should be run, especially if an asset has
it's attached file changed from an image to a non-image. Good enough.

Closes #313