DownloadDropdownDisplay no longer uses derivative models.
Instead it just creates blind links based on what derivatives it knows ought to exist for a given content type.

This means we can no longer supply file size of download to user with link (we would get that from deriv model). It also means we no longer check to make sure a derivative really exists before making download link -- if a user clicks on a download link for a deriv that doesn't actually exist, they'll just get a 404. These should be the only UI changes/degradatins.

This is an attempt at performance improvement. Accessing shrine json metadata (for file size, width, and height) seems to be a mysteriously expensive operation (may be fixed in shrine 2).

It also means that as far as DownloadDropdownDisplay is concerned, we COULD stop fetching/pre-loading derivative models at all. Which would be additional performance improvements. At present we are still preload fetching them for use as thumbnail src though. We will benchmark how much of an improvement we get from this step, without that.