Upload DZI files to S3 using concurrent threads
Using API from ruby-concurrent, which is already a rails dependency. http://ruby-concurrency.github.io/concurrent-ruby/master/file.promises.out.html

We use a custom thread pool to limit concurrency to 8 threads, to avoid accidentally using lots of resources or whatever. Experimentation showed more than 8 doesn't give much advantage.

Since there can be a LOT of tiles, using 8 concurrent threads to upload was observed to result in major performance improvement, as opposed to waiting for each to upload one by one serially. This works on MRI even with the GIL, becuase uploading is mostly waiting on IO, and MRI threads can work concurrently when busy on IO. The limit is probably when the network to S3 is saturated, and/or local disk IO is saturated.

There can be quite a few tiles.