avoid double display of same link to opac
When a bib number is entered as an identifier, PLUS a link to opac is included as a related_url (which happens in some of our older data), don't show the same link to opac twice. The solution is just putting a 'uniq' in the right place. With a unit test.

Ref #451