Lock to sprockets 3
Upgrading to sprockets 4 breaks our app until we tweak it to work. Putting this in Gemfile will keep any 'bundle update's of other dependencies from accidentally updating sprockets 4 too, before we're ready. When we're ready, we change or removee this line in Gemfile. We did create an issue to upgrade to sprockets 4 eventually, #458