Remove reference to outdated 'scih-uploads' bucket name
That is no longer a bucket we use. Don't try to guess what buckets we're using as a 'default' in our Env.rb, just count on local_env.yml to provide it, no reason to try to keep an additional source of information that has to be kept sync'd as a default in ruby code.

There is still a default for non-production environments, so when you're in dev you don't need to have a local_env.yml, we want dev to work out of the box. The dev default is a bucket not actually mounted anywhere.