btn-group class not actually necessary for bootstrap dropdown
And in at least some cases it was messing up popper positioning to avoid going out of bounds.

Bootstrap examples at https://getbootstrap.com/docs/4.0/components/dropdowns/ use 'btn-group' when needed for a divided button, as well as other times; as an alternative to 'dropdown' class. But we are using 'dropdown' class. Bootstrap examples are a bit confusing.