move custom label to i18n
The attribute name is called 'museum_object_id', but we realized that wasn't quite right, and what we want to show to user is 'Object ID (Past Perfect)'. The prior way of customizing this succesfully changed the label for hte input on screen via simple_form. But validation errors still said 'Museum Object can't be blank'. (Rails assumes 'id' on the end is an internal fk and ignores the 'id' part, but 'museum object id' isn't right anyway).

By changing with an i18n key conventional to Rails, now it's changed in field label, validation error message, etc.