Display 'required' badge next to collecting-area-specific required fields
Simple_form is smart enough to notice this from validations itself -- but not from conditional validations like our collecting-area-specific ones.

So we have to put our collecting-area-specific required fields in a data structure that rails code can introspect upon, so we can decide when to show the 'required' badge. Which makes this code a little bit more confusing/abstract, but not too bad.