Make featured topics thumbs smaller
So they will fit on a smaller (~1100px wide) screen still 3 accross. Now that we have 9 images instead of 6, let's keep them in rows of 3 longer to avoid so many rows.

Also limit to at MOST 3 across, as smaller ones could go to 4, but that wrapped oddly with 9 images. (4,4,1 was not great).

Note that when it does wrap to 2, we still get a row of just one at the end, 2,2,2,1 -- but that's just how it goes with 9 images instead of 6.