fix accordion open off-screen in OHMS ToC
This had already been done/fixed in the ohms-search branch, but we need to "back port" it to be in master while the unfortunately long-running ohms-search is still under construction. We'll have to deal with merge conflict over in ohms-search branch.

This is a little bit janky, it first opens the accordion, then repositions it to be at the right place. But this seems to be the best we can do with the bootstrap accordion stuff. I did notice similar "open then reposition" even in ohms standard viewer and other ohms-like viewers, it seems to be a common problem -- our fixed navbar makes it even worse. This is the best we can do, and hopefully good enough, we should ask stakeholders to verify it resolves problem reported in #689