Lock csl-styles gem to 1.0.1.9 exactly
Subsequent release causes our human-readable citations to be output in a way we do not prefer. For now, just lock it to the version that works. csl-styles being a separate gem with just style info is nice, cause we don't have to lock the actual csl logic gems.

See https://github.com/inukshuk/csl-styles/issues/5