upgrade minimist yarn dependency
Github has been complaining that we should upgrade minimist. Tried a new technique to do so.

1. Find the minimist lines in the `yarn.lock` manually, and delete them.
2. run `yarn` again to regenerate, finding latest satisfying versions.

In this case, we had two versions of minimist in the project, only one of them was able to be updated, because one of the things in the dependency tree (`mkdrp`) insists on using minimist 0.0.8 exactly. :(

But this commit does what we can.