Handle either shrine3 or shrine3 file not found exception classes
Shrine 3 helpfully standardizes the exception raised from storage when a file is not found. But that is a backwards compat.

In order to safely and easily get ready for shrine 3, we're making some changes such that tests will pass and the app will work on either shrine 2 OR shrine 3, make our app ambidextrous, with some tiny easily comprehended non-dangerous changes.

This is one of them.