Rake tasks to check derivatives -- do they all exist in storage?
The "opposite" of derivative orphan checker. Orphan checker goes through all files on S3, and makes sure they are all referenced in DB.

This one goes through all derivatives referenced in DB, and makes sure they exist on S3.

Turns out it's a LOT faster to download the complete list of paths on S3, and then check that locally for each derivative in DB. Rather than try to check each derivative in DB with a seperate request to S3 to see if it exists. Like orders of magnitude faster, this way can complete in seconds.

So we have two tasks, scihist:derivatives:dump dumps to a local SDBM db. (file-based key-value store built into ruby stdlib). The scihist:derivatives:check checks against it. This does mean if derivatives are being modified concurrently, it could be wrong if the dumped DB file of S3 paths is not up to date. It only takes seconds to dump though, so that's good.