Add searchable_plain_text_transcript column to oral_history_content table
This will be used for a plain text transcript that can be indexed in solr to support serach over it. It will be indexed in solr only for works that do NOT have an OHMS xml transcript, this will only be used when that's not present (and staff UI prompt should probably say that).

We are commiting it now first, before writing features on top of it, to get it in there so various features can be written on top without waiting for each other.

Having this in the table does mean that every time Rails fetches a row from this table (say to display an oral history), it will fetch back this columnw ith a lot of content, which isn't actually needed except at indexing time. I dn't think that will cause a problem, for now that's just what we're doing.