index OHMS fulltext into solr
Into a new field we define in solr schema, 'searchable_fulltext'. We're trying with omitNorms and without storing it.

In dev, after making a solr/schema.xml change, I discovered you need to:

    ./bin/rake solr:stop
    ./bin/rake solr:clean
    ./bin/rake solr:restart
    ./bin/rake scihist:solr:reindex

In staging/prod, capistrano should take care of all but the reindex.

We can merge this and fulltext will start being INDEXED into solr, but it's not being used by anything yet, it's not being included in searches. Boosts of existing fields may have to be changed to make room for this one, that's for later work.