Process as HTML when importing searchable OH fulltext
If it really was HTML (very small fraction of imports), it should turn it into plaintext, both getting rid of HTML tags and also resolving any HTML character entities.

If it's not really HTML but is plaintext... mostly it should be okay, but things like '<???>' or  will be interpreted as html tags, and stripped.

We have a hard automatically detecting what is really HTML because of nature of corrupt and inconsistent data. After talking with Lee, we thought this was the best bet for this import.