Redirect corrupted facet params
When sharing a search results URL on facebook, facebook will erroneously alter the facet limit params. From eg `?f[facet_name][0]=value` to `?f[facet_name][0]=value`

This will result in a Blacklight error. When the facebook crawler crawls it for a preview, as well as when any user clicks on it. We try to catch these, fix them to what Blacklight wants, and REDIRECT to the correct one again with an HTTP 301 Moved Permanently. (Better than two URLs representing the same search, and we will verify Facebook will follow redirect).