Eliminate extra .thumb div in MemberImagePresentation
The MemberImagePresentation ViewModel previously rendered a

  <div class="thumb">
   <a ...>

I realized it could be easily compressed to a single element, that was an 'a' with class 'thumb'. With only tiny change to CSS.

This reduces the complexity of our DOM somewhat, somewhat compensating for the extra container we just added for CSS aspect ratio trick.

In some cases there is a thumb that isn't an a (the image missing placeholder), but that still works fine too.

There are potentially other opportunities for collapsing multiple nested divs, but none I identified that would have so little cost to comprehensibility/maintainability as this one. This one seems worthwhile.