update lodash yarn dependency
For security warning.

It's a bit tricky to figure out how to update "indirect" dependencies (not specifically mentioned in package.json) with yarn.

Here, we first REMOVED the lodash "resolution" in package.json, which was an earlier attempt to deal with updating indirect dependencies, which wasn't really a good solution.

Then we followed instructions at https://medium.com/@ayushya/upgrading-javascript-packages-deep-dependencies-using-yarn-8b5983d5fb6b

* Hand-edit the yarn.lock to remove the relevant lodash resolution
* re-run `yarn install`

Bingo, now we have lodash 4.17.19 which is what we wanted.