Remove abstract from WorkShowInfo component
Because for some uses, in some Oral History layouts, we're going to want to put something between the abstract and the rest of WorkShowInfo.

This does mean that the various places using WorkShowInfo presently now need to copy-paste identical placements of the abstract/description above it, including getting the div classes right etc. We could pull that out into it's own component, but that seems like overkill, this is okay.

Added a test to fail on the standard work page if description is entirely missing, which wasn't there before.