Tell s3 to use far-future HTTP cache headeres in public derivatives responses
Our derivatives are in a public bucket, and we use public S3 URLs with them. Eg:

`https://scihist-digicoll-production-derivatives.s3.amazonaws.com/8c180322-f939-4ed1-8c70-bd6763260936/thumb_standard/701668916a578f43019f50a5149fe8bb.jpeg`

The last component of that URL is a long random string, supplied by shrine. Especially when "namespaced" by the asset PK and derivative type, it is virtually impossible for the same URL to ever refer to a different image. If the image were re-generated, it gets a new URL with a new long random string.

This means a derivative URL is basically permanent, and can be cached indefinitely. By responding with far-future cache headers (as long a cache value as HTTP spec allows), browsers can cache these images and not re-load them, or some intermediate caches could too. If we ever use a CDN, it also tells most CDNs that these are cache-foreverable.

We can set some options when uploading the derivative files to S3, to tell S3 when it delivers them, it should use these cache headers.

We were already setting this for DZI tiles (thankfully, there are so many!), but now do it for derivative files too.

Whether the comma is followed by a space or not in `max-age=31536000, public` doesn't matter, both are legal, but we'll use spaces since that seems to be in more common usage and is perhaps easier to read.