specific config for puma in heroku
from https://devcenter.heroku.com/articles/deploying-rails-applications-with-the-puma-web-server

These files shouldn't effect existing non-heroku environments, we just have a Procfile telling heroku we want it to start our web dynos with puma, with a specific config file we've also added for this purpose.

Puma is told to use two processes each with 5 threads. So that should be 10 workers total -- the same number we currently have via passenger (although they are each running in an entirely separate process via passenger). A heroku dyno may have 3 or 4 virtual CPUs -- in MRI we'd need a separate process per CPU to take full advantage of them parallely, so this is not necessarily taking full advantage of our CPUs, but not clear if we can fit more into the RAM we have.

We only had two CPUs on our current manual web EC2 anyway, so this should actually probably be roughly the same amount of capacity as we have on our manual EC2? Not entirely certain.