Avoid double fetch of all work members
What had been intended as an optimization ended up a de-optimization!

No new tests, instead refactor with preserved behavior all tests still passing.

This is worth understanding what's going on, to avoid making a similar mistake (this one was my mistake!). Happy to discuss it if needed.

This was found in exploring why so much memory for heroku; this was PART of it, but isn't a solution or anything.

The #representative_member method was fetching ALL members, and then filtering through it in-memory to find the representative member. Originally the author (uh me) thought this wouldn't be a double-fetch, because #members would already be in memory. But the implementation of #member_list_for_display was changed so it didn't actually load #members. And then the result was we ended up fetching ALL members (could be hundreds of items) TWICE.  This was way WORSE than if we had just used the existing #representative association to fetch the SINGLE representative object.

While there would be a way to actually do it in one fetch -- fetch all members, then split them into "represnetative" and "all but representative" -- instead we do a smaller refactor to just use the #representative association. This is good enough! By trying to be clever, we opened ourselves up to a bug that made it worse. Let's not be clever.

Here we actually can more easily avoid two fetches, fine, we will. While most importantly still avoiding a second fetch of ALL members to filter in-memory to find representative.